requirements.txt
requests
pytest
pyYaml